For RHEL AI 1.3: Remove manual chunking and separate document_collection from taxonomy. Update llm judge eval to use csvs, jsonls and gpt-4o
