+++ .gitignore
@@ -1,5 +1,7 @@
 .env
 eval/llm_config.yaml
+eval/.milvus_llm_judge_eval.db.lock
+eval/milvus_llm_judge_eval.db
 
 # Byte-compiled / optimized / DLL files
 __pycache__/
