+++ data_preparation/.env.example
@@ -1,2 +1,2 @@
 INPUT_DIR=document_collection
-OUTPUT_DIR=output
+OUTPUT_DIR=converted_documents
