+++ eval/ground_truth/qna_sample.csv
@@ -1,0 +1,3 @@
+question,ground_truth
+"Who are you?","I am a Red Hat Instruct Model, an AI language model developed by Red Hat and IBM Research based on the granite-3.0-8b-base model."
+"What kinds of questions and answers should I write?","You should work with the subject matter experts to write questions that are relevant to the content and answers that are accurate."
