+++ eval/ground_truth/qna_sample.jsonl
@@ -1,0 +1,2 @@
+{"question":"Can I use a jsonl formatted ground truth file?","ground_truth":"Yes, you can use a jsonl formatted ground truth file with columns question and ground_truth"}
+{"question":"What is artificial intelligence?","ground_truth":"Artificial intelligence, often abbreviated as AI, refers to the simulation of human intelligence in machines that are programmed to think and learn like humans. These intelligent systems can perform tasks such as recognizing speech, making decisions, solving problems, and even playing games. There are various types of AI, ranging from narrow AI, which is designed to perform a specific task, to general AI, which would theoretically be capable of performing any intellectual task that a human can do."}
