+++ eval/requirements.txt
@@ -1,18 +1,14 @@
-docling~=2.4.0
-docling-core~=2.3.1
-docling-ibm-models~=2.0.3
-docling-parse~=2.0.3
+docling[tesserocr]~=2.8.3
 einops~=0.8.0
-langchain~=0.3.5 
-langchain-community~=0.3.3 
-langchain-openai~=0.2.4
-langchain-milvus~=0.1.6
-langchain-text-splitters~=0.3.0
-openai~=1.53.0
-pypdf~=5.1.0
-pymilvus~=2.4.9
+langchain~=0.3.11
+langchain-community~=0.3.11
+langchain-openai~=0.2.12
+langchain-milvus~=0.1.7
+langchain-text-splitters~=0.3.2
+openai~=1.57.3
+pymilvus~=2.5.0
 python-dotenv~=1.0.1
-sentence-transformers~=3.2.1
-torch~=2.5.1
-torchvision~=0.20.1
-ragas~=0.2.5
+sentence-transformers~=3.3.1
+torch
+torchvision
+ragas~=0.2.8
